More SceneMenu decompilation

+++ src/Game/Menu/SceneMenu.hpp
@@ -19,7 +19,7 @@ public:
 
     }
 
-    void fn_800077A8(u32);
+    void fn_800077A8(u8);
 
     static CExScene *create(u16 arg0);
     static void fn_80006FA4(void);
